Add basic interaction between clients and server
I just added some basic functionality that could be useful based on
what has been proposed so far.
When a 'player' enters it can click on a button and the 'shared-env'
client gets notified but after that the button is disabled so the
shared-env is not notified anymore for this user.